# docs/openapi/simulation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Forge Simulation Service",
    "description": "Jobs, simulation templates, case files, and runner orchestration (internal; browser uses BFF).",
    "version": "1.0.0",
    "x-forge-service": "simulation",
    "x-forge-label": "Simulation"
  },
  "paths": {
...
    "/api/simulations/{sim_id}/run": {
      "post": {
        "summary": "Run Simulation",
        "description": "Create a job from a simulation template, optionally overriding input fields.",
        "operationId": "simulation.run_simulation_api_simulations__sim_id__run_post",
        "parameters": [
          {
            "name": "sim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sim Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Simulation Api Simulations  Sim Id  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-forge-service": "simulation",
        "x-forge-original-path": "/api/simulations/{sim_id}/run"
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "List Jobs",
        "operationId": "simulation.list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          }
        },
        "x-forge-service": "simulation",
        "x-forge-original-path": "/api/jobs"
      },
      "post": {
        "summary": "Create Job",
        "operationId": "simulation.create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_api_jobs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Job Api Jobs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-forge-service": "simulation",
        "x-forge-original-path": "/api/jobs"
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "simulation.get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Job Api Jobs  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-forge-service": "simulation",
        "x-forge-original-path": "/api/jobs/{job_id}"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_job_api_jobs_post": {
        "properties": {
          "case_zip": {
            "type": "string",
            "format": "binary",
            "title": "Case Zip"
          },
          "commands": {
            "type": "string",
            "title": "Commands",
            "default": ""
          },
          "run_instruction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Instruction Id"
          }
        },
        "type": "object",
        "required": [
          "case_zip"
        ],
        "title": "Body_create_job_api_jobs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "Simulation"
    }
  ]
}